fix(timeline-web): review feedback
