+++ packages/pluggableWidgets/timeline-web/src/assets/lineAndDot.ts
